refactor: split commands per domain and move screen-size styling to css
