+++ src/app/components/IncomingCallModal.css.ts
@@ -1,0 +1,16 @@
+import { style } from '@vanilla-extract/css';
+import { config } from 'folds';
+import { MOBILE_BREAKPOINT } from '$hooks/useScreenSize';
+
+export const Content = style({
+  padding: config.space.S600,
+  paddingBottom: `max(${config.space.S600}, env(safe-area-inset-bottom))`,
+  gap: config.space.S500,
+  '@media': {
+    [`(max-width: ${MOBILE_BREAKPOINT}px)`]: {
+      padding: config.space.S400,
+      paddingBottom: `max(${config.space.S500}, env(safe-area-inset-bottom))`,
+      gap: config.space.S400,
+    },
+  },
+});
