+++ src/app/features/call-status/styles.css.ts
@@ -1,5 +1,6 @@
 import { style } from '@vanilla-extract/css';
 import { color, config, toRem } from 'folds';
+import { MOBILE_BREAKPOINT } from '../../hooks/useScreenSize';
 
 export const LiveChipText = style({
   color: color.Critical.Main,
@@ -9,6 +10,14 @@ export const CallStatus = style([
   {
     padding: `${toRem(6)} ${config.space.S200}`,
     borderTop: `${config.borderWidth.B300} solid ${color.Background.ContainerLine}`,
+    flexDirection: 'row',
+    alignItems: 'center',
+    '@media': {
+      [`(max-width: ${MOBILE_BREAKPOINT}px)`]: {
+        flexDirection: 'column',
+        alignItems: 'normal',
+      },
+    },
   },
 ]);
