+++ src/app/features/room/RoomViewHeader.css.ts
@@ -1,5 +1,14 @@
 import { style } from '@vanilla-extract/css';
 import { config } from 'folds';
+import { MOBILE_BREAKPOINT } from '$hooks/useScreenSize';
+
+export const HeaderBalance = style({
+  '@media': {
+    [`(max-width: ${MOBILE_BREAKPOINT}px)`]: {
+      paddingLeft: config.space.S200,
+    },
+  },
+});
 
 export const HeaderTopic = style({
   ':hover': {
