+++ src/app/hooks/commands/parsing.ts
@@ -1,0 +1,109 @@
+import type { Room, RoomMember } from '$types/matrix-sdk';
+import { isServerName, isUserId } from '$utils/matrix';
+import { splitWithSpace } from '$utils/common';
+
+const FLAG_PAT = String.raw`(?:^|\s)-(\w+)\b`;
+const FLAG_REG = new RegExp(FLAG_PAT);
+const FLAG_REG_G = new RegExp(FLAG_PAT, 'g');
+
+export const splitPayloadContentAndFlags = (payload: string): [string, string | undefined] => {
+  const flagMatch = new RegExp(FLAG_REG).exec(payload);
+
+  if (!flagMatch) {
+    return [payload, undefined];
+  }
+  const content = payload.slice(0, flagMatch.index);
+  const flags = payload.slice(flagMatch.index);
+
+  return [content, flags];
+};
+
+export const parseFlags = (flags: string | undefined): Record<string, string | undefined> => {
+  const result: Record<string, string> = {};
+  if (!flags) return result;
+
+  const matches: { key: string; index: number; match: string }[] = [];
+
+  for (let match = FLAG_REG_G.exec(flags); match !== null; match = FLAG_REG_G.exec(flags)) {
+    if (!match[1] || !match[0]) continue;
+    matches.push({ key: match[1], index: match.index, match: match[0] });
+  }
+
+  for (let i = 0; i < matches.length; i += 1) {
+    const current = matches[i];
+    if (!current) continue;
+    const { key, match } = current;
+    const start = current.index + match.length;
+    const end = i + 1 < matches.length ? (matches[i + 1]?.index ?? flags.length) : flags.length;
+    const value = flags.slice(start, end).trim();
+    result[key] = value;
+  }
+
+  return result;
+};
+
+export const parseUsers = (payload: string): string[] => {
+  const users: string[] = [];
+
+  splitWithSpace(payload).forEach((item) => {
+    if (isUserId(item)) {
+      users.push(item);
+    }
+  });
+
+  return users;
+};
+
+export const parseServers = (payload: string): string[] => {
+  const servers: string[] = [];
+
+  splitWithSpace(payload).forEach((item) => {
+    if (isServerName(item)) {
+      servers.push(item);
+    }
+  });
+
+  return servers;
+};
+
+export const getServerMembers = (room: Room, server: string): RoomMember[] => {
+  const members: RoomMember[] = room
+    .getMembers()
+    .filter((member) => member.userId.endsWith(`:${server}`));
+
+  return members;
+};
+
+export const parseTimestampFlag = (input: string): number | undefined => {
+  const match = input.match(/^(\d+(?:\.\d+)?)([dhms])$/); // supports floats like 1.5d
+
+  if (!match) {
+    return undefined;
+  }
+
+  const value = Number.parseFloat(match[1]!); // supports decimal values
+  const unit = match[2];
+
+  const now = Date.now(); // in milliseconds
+  let delta = 0;
+
+  switch (unit) {
+    case 'd':
+      delta = value * 24 * 60 * 60 * 1000;
+      break;
+    case 'h':
+      delta = value * 60 * 60 * 1000;
+      break;
+    case 'm':
+      delta = value * 60 * 1000;
+      break;
+    case 's':
+      delta = value * 1000;
+      break;
+    default:
+      return undefined;
+  }
+
+  const timestamp = now - delta;
+  return timestamp;
+};
